chore: change seed files for 2024 edition
